refactor content.js
